requirements.txt
accelerate==1.2.0
diffusers==0.31.0
gradio==6.8.0
huggingface_hub==0.36.2
insightface==0.7.3
numpy==2.4.2
opencv_python==4.12.0.88
opencv_python_headless==4.12.0.88
peft==0.14.0
Pillow==12.1.1
safetensors==0.7.0
scipy==1.17.1
torch==2.3.0+cu118
torchvision==0.18.0+cu118
tqdm==4.67.1
transformers==4.47.0